tests/e2e/requirements.txt
pytest==7.4.0
playwright==1.38.0
python-dotenv==1.0.0